Suggestions tweaks
