+++ lib/suggestion_box_web/controllers/suggestion_controller.ex
@@ -19,7 +19,7 @@ defmodule SuggestionBoxWeb.SuggestionController do
       {:ok, suggestion} ->
         conn
         |> put_flash(:info, "Suggestion created successfully.")
-        |> redirect(to: Routes.suggestion_path(conn, :show, suggestion))
+        |> redirect(to: Routes.page_path(conn, :index))
 
       {:error, %Ecto.Changeset{} = changeset} ->
         render(conn, "new.html", changeset: changeset)
