+++ lib/suggestion_box_web/templates/layout/app.html.eex
@@ -9,7 +9,7 @@
   </head>
   <body>
 
-    <main role="main" class="container">
+    <main role="main" class="container pt-5">
       <%= if get_flash(@conn, :info) do %>
         <p class="alert alert-info" role="alert"><%= get_flash(@conn, :info) %></p>
       <% end %>
