+++ lib/suggestion_box_web/templates/suggestion/index.html.eex
@@ -1,4 +1,4 @@
-<div class="jumbotron mt-5 text-center">
+<div class="jumbotron text-center">
   <h1>Manage Suggestions</h1>
 </div>
